+++ tests/tests.py
@@ -1105,3 +1105,21 @@ def test_filegraph():
         )
     # make sure the generated pdf file is not empty
     assert os.stat(pdf_path).st_size > 0
+
+
+def test_batch():
+    from snakemake.dag import Batch
+
+    run(dpath("test_batch"), batch=Batch("aggregate", 1, 2))
+
+
+def test_batch_final():
+    from snakemake.dag import Batch
+
+    run(dpath("test_batch_final"), batch=Batch("aggregate", 1, 1))
+
+
+def test_batch_fail():
+    from snakemake.dag import Batch
+
+    run(dpath("test_batch"), batch=Batch("aggregate", 2, 2), shouldfail=True)